+++ Magic/CHANGELOG.md
@@ -22,6 +22,7 @@
  - Add Interval action, like a combo of Repeat + Delay
  - Fix overlapping casts overwriting each others' variables
  - Fix blocked-spells and allowed-spells flags (broken in 7.9.2)
+ - The SpawnEntity action now forces spawned mobs to ignore their owner by default. Set "owned: false" to prevent this.
  
 # 7.9.2